Add on-change listener for stories dropdown
When the user selects a story from the dropdown (above the Compose window), the story text is inserted in the compose window.

I initially wrote the select dropdown using ng-repeat (see below), but then realized I could use ng-options instead, which seems to give me more control. I'm not actually using that control yet, but it's also a bit more elegant a solution.

          <select ng-model="selectedStory" name="message--stories" id="message--stories">
            <option value="null">Select a story to include...</option>
            <option
                ng-repeat="story in stories | filter:contactStoryMatcher(contact)"
                ng-click="pasteStory(contact, story)"
                value="{{ story.title }}">
              {{ story.title }}
            </option>
          </select>